requirements.txt
# G spread
gspread==3.4.2
oauth2client==4.1.3

# Mailchimp
mailchimp3==3.0.13

# Firebase
python-firebase~=1.2

# Sound
Pygame==1.9.6

#Images
Pillow==7.1.1

future~=0.18.2

setuptools~=46.1.3
firebase~=3.0.1